refact: cria componentes para separar o código
