+++ src/pages/Home/Countdown/index.tsx
@@ -1,0 +1,13 @@
+import { CountdownContainer, Separator } from './styles'
+
+export function Countdown() {
+  return (
+    <CountdownContainer>
+      <span>{minutes[0]}</span>
+      <span>{minutes[1]}</span>
+      <Separator>:</Separator>
+      <span>{seconds[0]}</span>
+      <span>{seconds[1]}</span>
+    </CountdownContainer>
+  )
+}
