+++ src/pages/Home/NewCycleForm/index.tsx
@@ -1,0 +1,36 @@
+import { FormContainer, MinutesAmountInput, TaskInput } from './styles'
+
+export function NewCycleForm() {
+  return (
+    <FormContainer>
+      <label htmlFor="task">Vou trabalhar em</label>
+      <TaskInput
+        id="task"
+        list="task-suggestions"
+        placeholder="Dê um nome para o seu projeto"
+        disabled={!!activeCycle}
+        {...register('task')}
+      />
+
+      <datalist id="task-suggestions">
+        <option value="Tarefa 1" />
+        <option value="Tarefa 2" />
+        <option value="Tarefa 3" />
+      </datalist>
+
+      <label htmlFor="minutesAmount">durante</label>
+      <MinutesAmountInput
+        type="number"
+        id="minutesAmount"
+        placeholder="00"
+        step={5}
+        min={1}
+        max={60}
+        disabled={!!activeCycle}
+        {...register('minutesAmount', { valueAsNumber: true })}
+      />
+
+      <span>minutos</span>
+    </FormContainer>
+  )
+}
